Move extension into separate file
